Preserve 64-bit integers on 32-bit platforms
This commit ensures the preservation of 64-bit integers on 32-bit
platforms by replacing Int with Int64 as the basic, raw binding type.

Int (like Bool before it) is supported in the expression layer as a
Value type. The one part of the query layer that has changed is that
rowids are now emitted as Int64, which means you should probably update
your expressions:

    let id = Expression<Int64>("id")

Because integer literals convert to Double where they can (and not
Int64), there were some confusing results in the raw SQL layer:

    db.scalar("SELECT * FROM users WHERE id = ?", 1)
    // SELECT * FROM users WHERE id = 1.0

To avoid this (and make binding simpler), both Int and Bool now conform
to Bindable.

    db.run("SELECT * FROM users WHERE admin = ?", false)

When a row is unpacked into an array of bindings or scalar, however,
INTEGER values will always become Int64. Thus some gotchas:

    for row in stmt {
        row[0] as! Int // crashes, even if an integer
    }
    stmt.scalar() as? Int // silently fails, even if an integer

Cursors, meanwhile, can return Bool and Int values just fine.

    while stmt.cursor.step() {
        stmt.cursor[0] as Int
    }

+++ SQLite Tests/DatabaseTests.swift
@@ -66,10 +66,10 @@ class DatabaseTests: XCTestCase {
     }
 
     func test_scalar_preparesRunsAndReturnsScalarValues() {
-        XCTAssertEqual(0, db.scalar("SELECT count(*) FROM users WHERE admin = 0") as Int)
-        XCTAssertEqual(0, db.scalar("SELECT count(*) FROM users WHERE admin = ?", 0) as Int)
-        XCTAssertEqual(0, db.scalar("SELECT count(*) FROM users WHERE admin = ?", [0]) as Int)
-        XCTAssertEqual(0, db.scalar("SELECT count(*) FROM users WHERE admin = $admin", ["$admin": 0]) as Int)
+        XCTAssertEqual(Int64(0), db.scalar("SELECT count(*) FROM users WHERE admin = 0") as Int64)
+        XCTAssertEqual(Int64(0), db.scalar("SELECT count(*) FROM users WHERE admin = ?", 0) as Int64)
+        XCTAssertEqual(Int64(0), db.scalar("SELECT count(*) FROM users WHERE admin = ?", [0]) as Int64)
+        XCTAssertEqual(Int64(0), db.scalar("SELECT count(*) FROM users WHERE admin = $admin", ["$admin": 0]) as Int64)
     }
 
     func test_transaction_beginsAndCommitsStatements() {
@@ -217,14 +217,14 @@ class DatabaseTests: XCTestCase {
         db.create(collation: "NODIACRITIC") { lhs, rhs in
             return lhs.compare(rhs, options: .DiacriticInsensitiveSearch)
         }
-        XCTAssertEqual(1, db.scalar("SELECT ? = ? COLLATE NODIACRITIC", "cafe", "café") as Int)
+        XCTAssertEqual(Int64(1), db.scalar("SELECT ? = ? COLLATE NODIACRITIC", "cafe", "café") as Int64)
     }
 
     func test_createCollation_createsQuotableCollation() {
         db.create(collation: "NO DIACRITIC") { lhs, rhs in
             return lhs.compare(rhs, options: .DiacriticInsensitiveSearch)
         }
-        XCTAssertEqual(1, db.scalar("SELECT ? = ? COLLATE \"NO DIACRITIC\"", "cafe", "café") as Int)
+        XCTAssertEqual(Int64(1), db.scalar("SELECT ? = ? COLLATE \"NO DIACRITIC\"", "cafe", "café") as Int64)
     }
 
 }
